requirements.txt
telebot
instaloader
python-dotenv
requests[socks]